fix(ui): address PR review — conditionCount, AUDIT-001/003, ClassicWorkflow warning
- BusinessRulesList: conditionCount now sums across all conditionGroups
  (was reading only [0].conditions.length, undercounting ELSE IF rules)
- ClassicWorkflowProcessor: catch block now pushes a non-partial StepWarning
  with the error message instead of silently returning []
- PipelineTracesPanel: move ENTITY_COLORS palette to CrossEntityAutomationView
  (the AUDIT-003 approved exception file); pass entityColorMap as a prop
- FieldPills: mpillHit backgroundColor changed from colorPaletteGreenBackground2
  to colorNeutralBackground1 (AUDIT-001 — palette bg forbidden on text spans)
- FieldMatchVerdict: fieldMatchFires/fieldMatchNoFilter switched from palette
  backgrounds to neutral bg + coloured left-border accent (AUDIT-001)
