+++ src/components/CrossEntityAutomationView.tsx
@@ -1,4 +1,4 @@
-import { useState } from 'react';
+import { useMemo, useState } from 'react';
 import {
   Text,
   Title3,
@@ -17,6 +17,32 @@ import { SummaryStatsGrid } from './crossEntity/SummaryStatsGrid';
 import { RiskWarningsSection } from './crossEntity/RiskWarningsSection';
 import { PipelineTracesPanel } from './crossEntity/PipelineTracesPanel';
 import { GlobalChainMapPanel } from './crossEntity/GlobalChainMapPanel';
+
+/* ─────────────────────────────────────────────────────────────────────────
+   Entity accent colour palette — cycles per entity in display order.
+   AUDIT-003 exception: intentional hardcoded hex values — Fluent UI tokens
+   do not provide a cycling multi-entity palette. Values match Fluent UI brand
+   colours: blue=colorBrandBackground, green=colorPaletteGreenBackground2,
+   orange=colorPaletteMarigoldBackground2, purple=colorPaletteVioletBackground2,
+   teal=colorPaletteTealBackground2, pink=colorPaletteMagentaBackground2,
+   hot pink=colorPaletteHotPinkBackground2, dark teal=colorPaletteDarkGreenBackground2
+───────────────────────────────────────────────────────────────────────── */
+const ENTITY_COLORS = [
+  '#0078d4', '#107c10', '#ca5010', '#8764b8',
+  '#038387', '#c239b3', '#e3008c', '#004b50',
+];
+
+function buildEntityColorMap(analysis: CrossEntityAnalysisResult): Map<string, string> {
+  return new Map(
+    Array.from(analysis.allEntityPipelines.keys())
+      .sort((a, b) => {
+        const pa = analysis.allEntityPipelines.get(a)!;
+        const pb = analysis.allEntityPipelines.get(b)!;
+        return pa.entityDisplayName.localeCompare(pb.entityDisplayName);
+      })
+      .map((key, i) => [key, ENTITY_COLORS[i % ENTITY_COLORS.length]])
+  );
+}
 
 const useStyles = makeStyles({
   container: { display: 'flex', flexDirection: 'column', gap: tokens.spacingVerticalL },
@@ -42,6 +68,7 @@ export function CrossEntityAutomationView({
 }: CrossEntityAutomationViewProps): JSX.Element {
   const styles = useStyles();
   const [subView, setSubView] = useState<string>('traces');
+  const entityColorMap = useMemo(() => analysis ? buildEntityColorMap(analysis) : new Map<string, string>(), [analysis]);
 
   if (!analysis || (analysis.allEntityPipelines.size === 0 && analysis.totalEntryPoints === 0)) {
     return (
@@ -76,7 +103,7 @@ export function CrossEntityAutomationView({
         <Tab value="map">Global Chain Map ({analysis.chainLinks.length})</Tab>
       </TabList>
 
-      {subView === 'traces' && <PipelineTracesPanel analysis={analysis} />}
+      {subView === 'traces' && <PipelineTracesPanel analysis={analysis} entityColorMap={entityColorMap} />}
 
       {subView === 'map' && <GlobalChainMapPanel analysis={analysis} />}
     </div>
